Advance the MEOS commit and widen the index row identifier (#288)
* Advance the MEOS commit and widen the index row identifier

The recorded MEOS commit moves to 83f22a321a70, where the in-memory index
takes and returns a row identifier as int64 rather than int: rtree_insert,
rtree_insert_temporal, rtree_insert_temporal_split, rtree_nn_cursor_next and
their four space-partitioning counterparts. Those eight signatures are the
whole public-header difference between the two commits.

The index modules carry that identifier at the old width in twelve places,
and only two of them fail to compile -- `int id` handed to the cursor. The
silent ten are the result array, created with sizeof(int) while the tree
writes int64 elements into it and read back through an `*(int *)` cast, and
the six inserts, where a static_cast<int> narrows a row_t before passing it
to a parameter that is now int64: a row identifier above 2^31 would be stored,
and later searched for, as a different number.

The regenerated surface belongs to the same commit because a surface derived
from one MEOS commit and a library built from another are not one artifact.
It registers eleven further nearest-approach overloads, three of them against
a temporal box, which is what makes the T_TBOX arm of the matcher and of the
exact-distance dispatch reachable from SQL.

* Cover the nearest-neighbour path over a temporal box

The nearest-neighbour fixtures index a spatiotemporal box. An index over a
temporal number stores a temporal box instead, so `|=|` against one reaches a
different arm of both the matcher and the exact-distance dispatch, and that
arm has no fixture behind it.

The temporal float and the temporal integer are covered separately because a
bare TBOX literal carries a float span whatever its bounds look like, so a
temporal integer needs the TBOXINT form; the two span types are two value
domains and comparing across them is an error rather than a number, which the
test states. Both index kinds answer the query, and the declined shapes -- a
pushed-down filter and a descending order -- are checked on this arm too.

---------

+++ src/index/rtree_module.cpp
@@ -247,7 +247,7 @@ void TRTreeIndex::ReplayEntries() {
                 continue;
             }
             memcpy(box.data(), entry + sizeof(row_t), bbox_size_);
-            rtree_insert(rtree_, box.data(), static_cast<int>(row_id));
+            rtree_insert(rtree_, box.data(), row_id);
         }
 
         entry_count_ += count;
@@ -467,7 +467,7 @@ void TRTreeIndex::Construct(DataChunk &expression_result, Vector &row_identifier
             }
         }
 
-        rtree_insert(rtree_, box.data(), static_cast<int>(row_data[i]));
+        rtree_insert(rtree_, box.data(), row_data[i]);
         RecordEntry(box.data(), row_data[i]);
     }
 }
@@ -479,7 +479,7 @@ ErrorData TRTreeIndex::BulkConstruct(STBox* boxes, const row_t* row_ids, idx_t c
     }
 
     for (idx_t i = 0; i < count; i++) {
-        rtree_insert(rtree_, &boxes[i], static_cast<int>(row_ids[i]));
+        rtree_insert(rtree_, &boxes[i], row_ids[i]);
         RecordEntry(&boxes[i], row_ids[i]);
     }
 
@@ -662,7 +662,7 @@ bool TRTreeIndex::NNScanNext(IndexScanState &state, row_t &row_id, double &lower
     if (!nstate.cursor) {
         return false;
     }
-    int id = 0;
+    int64 id = 0;
     double distance = 0;
     if (!rtree_nn_cursor_next(nstate.cursor, &id, &distance)) {
         return false;
@@ -679,7 +679,7 @@ vector<row_t> TRTreeIndex::Search(const void *query_box, RTreeSearchOp op) const
         return results;
     }
 
-    MeosArray *ids = meos_array_create(sizeof(int));
+    MeosArray *ids = meos_array_create(sizeof(int64));
 
     try {
         int count = rtree_search(rtree_, op, query_box, ids);
@@ -687,7 +687,7 @@ vector<row_t> TRTreeIndex::Search(const void *query_box, RTreeSearchOp op) const
         if (count > 0) {
             results.reserve(count);
             for (int i = 0; i < count; i++) {
-                const auto row_id = static_cast<row_t>(*(int *) meos_array_get(ids, i));
+                const auto row_id = static_cast<row_t>(*(int64 *) meos_array_get(ids, i));
                 // Deleted rows are still in the tree; they are filtered here.
                 if (deleted_.find(row_id) != deleted_.end()) {
                     continue;
